+++ src/components/List/GameInfoList.vue
